Refactor `kmersim` command to add ANI estimation
If the 'jaccard' scoring metric is enabled, automatically add the
estimated ANI too.

This commit also refactors the multiprocessing code a bit. First,
we loaded the kmerset before hand, and cached every k-merset in the
main process. I've changed that to caching the k-merset on a per
process basis. This makes the code a bit cleaner, we have access to
the full k-merset details (like the actual value of `k` used), and
because we use a large enough chunksize, I expect not too much of a
performance hit.